factor out commong file writing code
