+++ lib/chisel-utils.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const indent = require('./indent.js');
-const paramSchema = require('./param-schema.js');
 
 const toChiselType = (width) => {
   return (width === 1 ? 'Bool()' : 'UInt((' + width + ').W)');
